Casts max atoms variable to int, CRLF to LF

+++ include/Defs.h
@@ -2,7 +2,7 @@
 
 // The definitions in this file primarily relate specifically to the simulation module
 // TODO: change to heap memory so not limited by stack
-#define MAXIMUM_NUMBER_OF_ATOMS 5e5					// maximum number of atoms; each atom is allocated dynamically
+#define MAXIMUM_NUMBER_OF_ATOMS (int) 5e5					// maximum number of atoms; each atom is allocated dynamically
 //#define MAXIMUM_NUMBER_OF_SURFACE_ATOMS 2000000		// TODO: move this to input file or calculated from system size
 
 #define MAXIMUM_NUMBER_OF_ACTIVATION_BARRIERS 100		//used! // maximum number of distinct activation barriers for different
